@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "cordon-blue-front",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,38 +8,36 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "alertifyjs": "^1.13.1",
-    "axios": "^0.19.2",
-    "bootstrap-vue": "^2.23.1",
-    "core-js": "^3.21.1",
-    "deep-extend": "^0.5.1",
-    "vue": "^2.7.14",
-    "vue-alertify": "^1.1.0",
-    "vue-bootstrap-datetimepicker": "^5.0.1",
-    "vue-router": "^3.5.3",
-    "vuejs-datepicker": "^1.6.2",
-    "vuex": "^3.6.2"
+    "core-js": "^3.8.3",
+    "vue": "^3.2.13"
   },
   "devDependencies": {
-    "@babel/polyfill": "^7.12.1",
-    "@vue/cli-plugin-babel": "^4.5.15",
-    "@vue/cli-plugin-router": "^4.5.15",
-    "@vue/cli-plugin-vuex": "^4.5.15",
-    "@vue/cli-service": "^4.5.15",
-    "babel-eslint": "^10.1.0",
-    "bootstrap": "^4.6.2",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "mutationobserver-shim": "^0.3.7",
-    "popper.js": "^1.16.0",
-    "portal-vue": "^2.1.6",
-    "sass": "^1.49.8",
-    "sass-loader": "^8.0.0",
-    "vue-cli-plugin-bootstrap-vue": "^0.6.0",
-    "vue-template-compiler": "^2.6.14"
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "@babel/eslint-parser"
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
-    "last 2 versions"
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
   ]
 }
