@@ -1,16 +1,4 @@
-import "@babel/polyfill";
-import "mutationobserver-shim";
-import Vue from "vue";
-import "./plugins/bootstrap-vue";
-import App from "./App.vue";
-import store from "./store/store.js";
-import VueAlertify from "vue-alertify";
+import { createApp } from 'vue'
+import App from './App.vue'
 
-Vue.use(VueAlertify);
-
-Vue.config.productionTip = false;
-
-new Vue({
-  store,
-  render: h => h(App)
-}).$mount("#app");
+createApp(App).mount('#app')
