@@ -1,34 +1,26 @@
 <template>
-  <div id="app">
-    <filter-component />
-   <!-- <Navbar />
-   <NewEventComponent /> -->
-  </div>
+  <img alt="Vue logo" src="./assets/logo.png">
+  <HelloWorld msg="Welcome to Your Vue.js App"/>
 </template>
 
 <script>
-import FilterComponent from "./components/FilterComponent.vue";
-import NewEventComponent from './components/NewEventComponent.vue';
-import Navbar from './components/Navbar.vue';
+import HelloWorld from './components/HelloWorld.vue'
 
 export default {
-  name: "App",
+  name: 'App',
   components: {
-    FilterComponent,
-    NewEventComponent,
-    Navbar,
-  },
-};
+    HelloWorld
+  }
+}
 </script>
 
 <style>
-@import url("https://fonts.googleapis.com/css?family=Heebo");
-
 #app {
-  font-family: "Heebo";
+  font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
+  margin-top: 60px;
 }
 </style>
